Refactor name and directory
클래스명 변경, 패키지 디렉토리 정리
